refactor: split snail-codegen lib.rs into smaller modules
Split the large lib.rs file (1641 lines) into four focused modules:
- feature_detection.rs: Feature usage detection (try, regex, subprocess, structured accessor)
- helpers.rs: Helper code generation (try, regex, subprocess, structured accessor, jmespath)
- writer.rs: PythonWriter struct with statement generation
- expr.rs: Expression and AST node formatting functions

The main lib.rs now contains only the public API (python_source functions)
and module coordination. This improves code organization and maintainability.

All CI checks pass:
- cargo fmt --check: ✓
- RUSTFLAGS="-D warnings" cargo build: ✓
- cargo clippy -- -D warnings: ✓
- RUSTFLAGS="-D warnings" cargo test: ✓ (141 tests pass)
